backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.20.0
Pillow>=9.0.0
torch>=1.10.0
torchvision>=0.11.0